+++ test/pay/billable/stripe_test.rb
@@ -4,10 +4,7 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   setup do
     @billable = User.new email: "johnny@appleseed.com"
     @billable.processor = "stripe"
-
     @customer = ::Stripe::Customer.create(email: @billable.email)
-
-    @plan = ::Stripe::Plan.retrieve("small-monthly")
   end
 
   test "getting a stripe customer with a processor id" do
@@ -17,18 +14,15 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
 
   test "getting a stripe customer without a processor id" do
     assert_nil @billable.processor_id
-
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     @billable.customer
-
     assert_not_nil @billable.processor_id
     assert @billable.card_type == "Visa"
     assert @billable.card_last4 == "4242"
   end
 
   test "can create a charge" do
-    @billable.card_token = payment_method.id
-
+    @billable.card_token = payment_method
     charge = @billable.charge(2900)
     assert_equal Pay::Charge, charge.class
     assert_equal 2900, charge.amount
@@ -41,14 +35,14 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
 
   test "raises action required error when SCA required" do
     exception = assert_raises(Pay::ActionRequired) {
-      @billable.card_token = sca_payment_method.id
+      @billable.card_token = sca_payment_method
       @billable.charge(2900)
     }
     assert_equal "This payment attempt failed because additional action is required before it can be completed.", exception.message
   end
 
   test "can create a subscription" do
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     @billable.subscribe(name: "default", plan: "small-monthly")
 
     assert @billable.subscribed?
@@ -57,7 +51,7 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   end
 
   test "can swap a subscription" do
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     subscription = @billable.subscribe(name: "default", plan: "small-monthly")
     subscription.swap("small-annual")
     assert @billable.subscribed?
@@ -74,7 +68,7 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
 
   test "fails when subscribing with SCA card" do
     exception = assert_raises(Pay::ActionRequired) {
-      @billable.card_token = sca_payment_method.id
+      @billable.card_token = sca_payment_method
       @billable.subscribe(name: "default", plan: "small-monthly")
     }
 
@@ -82,29 +76,25 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   end
 
   test "can update their card" do
-    @billable.update_card(payment_method.id)
+    @billable.update_card payment_method
 
     assert_equal "Visa", @billable.card_type
-    assert_equal "4242", @billable.card_last4
     assert_nil @billable.card_token
 
-    payment_method = create_payment_method(card: {number: "6011 1111 1111 1117"})
-    @billable.update_card(payment_method.id)
-
+    @billable.update_card "pm_card_discover"
     assert @billable.card_type == "Discover"
-    assert @billable.card_last4 == "1117"
   end
 
   test "retriving a stripe subscription" do
     @billable.processor_id = @customer.id
-    @billable.update_card(payment_method.id)
+    @billable.update_card(payment_method)
     subscription = ::Stripe::Subscription.create(plan: "small-monthly", customer: @customer.id)
     assert_equal @billable.processor_subscription(subscription.id), subscription
   end
 
   test "can create an invoice" do
     @billable.processor_id = @customer.id
-    @billable.update_card(payment_method.id)
+    @billable.update_card(payment_method)
 
     ::Stripe::InvoiceItem.create(
       customer: @customer.id,
@@ -122,12 +112,9 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
     @billable.processor = "stripe"
     @billable.processor_id = @customer.id
 
-    assert_equal @billable.customer, @customer
-
-    @billable.card_token = payment_method.id
-
     # This should trigger update_card
-    assert_equal @billable.customer, @customer
+    @billable.card_token = payment_method
+    @billable.customer
     assert_equal @billable.card_type, "Visa"
     assert_equal @billable.card_last4, "4242"
   end
@@ -169,14 +156,14 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   end
 
   test "handles coupons" do
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     subscription = @billable.subscribe(plan: "small-monthly", coupon: "10BUCKS")
     assert_equal "10BUCKS", subscription.processor_subscription.discount.coupon.id
   end
 
   test "stripe trial period options" do
     travel_to(VCR.current_cassette.originally_recorded_at || Time.current) do
-      @billable.card_token = payment_method.id
+      @billable.card_token = payment_method
       subscription = @billable.subscribe(plan: "small-monthly", trial_period_days: 15)
       assert_equal "trialing", subscription.status
       assert_not_nil subscription.trial_ends_at
@@ -191,7 +178,7 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   end
 
   test "can pass shipping information to charge" do
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     charge = @billable.charge(25_00, shipping: {
       name: "Recipient",
       address: {
@@ -205,20 +192,29 @@ class Pay::Stripe::Billable::Test < ActiveSupport::TestCase
   end
 
   test "allows subscription quantities" do
-    @billable.card_token = payment_method.id
+    @billable.card_token = payment_method
     subscription = @billable.subscribe(plan: "small-monthly", quantity: 10)
     assert_equal 10, subscription.processor_subscription.quantity
     assert_equal 10, subscription.quantity
   end
 
+  test "card is automatically updated on subscribe" do
+    assert_nil @billable.card_type
+    @billable.update_card "pm_card_chargeCustomerFail"
+    assert_not_nil @billable.card_type
+    @billable.card_token = "pm_card_amex"
+    @billable.subscribe
+    assert_equal @billable.card_type, "Amex"
+  end
+
   private
 
   def payment_method
-    @payment_method ||= create_payment_method
+    @payment_method ||= "pm_card_visa"
   end
 
   def sca_payment_method
-    @sca_payment_method ||= create_payment_method(card: {number: "4000 0027 6000 3184"})
+    @sca_payment_method ||= "pm_card_authenticationRequired"
   end
 
   def create_payment_method(options = {})